Relocate RocketBoosterSafety()
Move the RocketBoosterSafety function from mis_c, which is primarily
drawing/animation, to the hardware utilities.
