+++ src/game/game_main.cpp
@@ -43,10 +43,10 @@
 #include "aimast.h"
 #include "ast4.h"
 #include "endgame.h"
+#include "hardware.h"
 #include "intel.h"
 #include "intro.h"
 #include "mc.h"
-#include "mis_c.h"
 #include "mission_util.h"
 #include "museum.h"
 #include "newmis.h"
