+++ src/game/newmis.cpp
@@ -27,8 +27,8 @@
 #include "game_main.h"
 #include "downgrader.h"
 #include "draw.h"
+#include "hardware.h"
 #include "ioexception.h"
-#include "mis_c.h"
 #include "place.h"
 #include "port.h"
 #include "state_utils.h"
