+++ src/game/vab.cpp
@@ -47,8 +47,8 @@
 #include "options.h"
 #include "admin.h"
 #include "game_main.h"
+#include "hardware.h"
 #include "logging.h"
-#include "mis_c.h"
 #include "mission_util.h"
 #include "news_suq.h"
 #include "place.h"
