+++ src/game/vehicle.cpp
@@ -26,7 +26,6 @@
 #include "data.h"
 #include "game_main.h"
 #include "hardware.h"
-#include "mis_c.h"
 
 
 /**
